refactor(30746): refactor the southbound mapping editor to extract the instruction list
